+++ config/expert/Xorandor/code/CGCode.cpp
@@ -1,0 +1,22 @@
+#include <iostream>
+#include <string>
+#include <vector>
+#include <algorithm>
+
+using namespace std;
+
+int main()
+{
+    int height;
+    int width;
+    cin >> height >> width; cin.ignore();
+    for (int i = 0; i < height; i++) {
+        string line;
+        getline(cin, line);
+    }
+
+    // Write an answer using cout. DON'T FORGET THE "<< endl"
+    // To debug: cerr << "Debug messages..." << endl;
+
+    cout << "anwser" << endl;
+}