+++ config/expert/Xorandor/code/CGCode.java
@@ -1,0 +1,23 @@
+import java.util.*;
+import java.io.*;
+import java.math.*;
+
+class Solution {
+
+    public static void main(String args[]) {
+        Scanner in = new Scanner(System.in);
+        int height = in.nextInt();
+        int width = in.nextInt();
+        if (in.hasNextLine()) {
+            in.nextLine();
+        }
+        for (int i = 0; i < height; i++) {
+            String line = in.nextLine();
+        }
+
+        // Write an answer using System.out.println()
+        // To debug: System.err.println("Debug messages...");
+
+        System.out.println("anwser");
+    }
+}