+++ config/expert/Xorandor/code/CGCode.ts
@@ -1,0 +1,11 @@
+var inputs: string[] = readline().split(' ');
+const height: number = parseInt(inputs[0]);
+const width: number = parseInt(inputs[1]);
+for (let i = 0; i < height; i++) {
+    const line: string = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('anwser');